refactor(constants): 관심사별 분류
