+++ src/pages/Home/index.tsx
@@ -3,8 +3,8 @@ import styled from '@emotion/styled';
 import Footer from '@/components/layouts/Footer';
 import { HEADER_HEIGHT } from '@/components/layouts/Header';
 import { TABBAR_HEIGHT } from '@/components/layouts/TabBar';
-import { homeAdList } from '@/constants/homeAdList';
-import { articleList } from '@/constants/articleList';
+import { homeAdList } from '@/constants/home/homeAdList';
+import { articleList } from '@/constants/home/articleList';
 import AdBanner from './AdBanner';
 import ArticleBanner from './ArticleBanner';
