+++ src/pages/Login/index.tsx
@@ -7,7 +7,7 @@ import Logo from '@/assets/logo.svg?react';
 import IconButton from '@/components/common/IconButton';
 import Header, { HEADER_HEIGHT } from '@/components/layouts/Header';
 import { RouterPath } from '@/routes/path';
-import { backgroundImageList } from '@/constants/backgroundImageList';
+import { backgroundImageList } from '@/constants/login/backgroundImageList';
 
 const Login = () => {
   const isMember = false; // 추후 API 연동
